Refactor utils/generate_contributors.py -> ssg/_contributors.py
 - Moves bulk of functionality into ssg/_contributors.py
 - Fixes python unit tests against generate_contributors
